Refactor configuration to support per-world settings.
Both player and claim data are now saved together with each world's
respective data.

For example, if the main world is named "world", GriefPrevention data
would be stored in the following location :

./world/GriefPreventionData/ClaimData/..
./world/GriefPreventionData/PlayerData/..

Configurations are still stored within ./config but is now "world-aware".
To have config settings only affect the main world, the following file
would be need to be modified :

./config/GriefPrevention/worlds/overworld/world/world.conf

* Claim ID's now use UUID's instead of integers.
* PlayerData is now world aware.