Add FlagPermissions for claim contexts.
Flag permissions will grant an admin to deny/allow protection/command use to a
player or group using a permissions plugin such as PermissionsEx. In order to
apply these permissions, you need to use the addflagpermissions command.
For example, if I wanted to set the block-place permission on a group
called 'admin', I would do the following :

/gp addflagpermissions group admin block-place true