Add item ban support.
* Add '/gp banitem <id>' command to ban items. This command requires the
permission 'griefprevention.command.banitem'.
* Add '/gp unbanitem <id>' command to unban items. This command requires the
permission 'griefprevention.command.unbanitem'.

Note: If no id is specified above, the item in hand will be banned or
unbanned. The format of id is "modid:name" or "modid:name:meta". The modid
used for vanilla items is "minecraft" so if you want to ban chest it would
be "minecraft:chest" or you can simply equip a chest in hand to ban it.

* Add permission 'griefprevention.ignorebans' to ignore all item bans.
This is useful for admins.