Add new flag 'item-spawn' for handling entity item spawns. Fixes #189
This new flag will not handle spawns from players such as a drop. It is
recommended to keep this flag as false for basic/admin claims or you can
run into dupes if an outside player attempts to break a block.

To allow items to spawn, either add trust or set the flag to true.