Add clickable teleport support for claim corners to /claiminfo.
Upon typing /claiminfo, you will now see 2 new lines : NorthCorners and
SouthCorners. NorthCorners represent the NW and NE corners of your claim
while SouthCorners represent the SW and SE corners of your claim.

To teleport to a corner, simply click anywhere in the text area. If the
location is safe, you will be instantly teleported there. Otherwise, if
the location is unsafe, you will be prompted and can force the teleport.

This should make it alot easier to manage larger claims.