Fix changed default values not being displayed in /cf command.
This fixes the issue with an admin changing the default value of a flag
and not being able to see it update when using the /cf command.

For example,

/cf block-break any true default

If standing in a basic claim, this would update the default 'block-break'
flag value to always default to true when a user creates a new basic
claim. However, this differs from the 'override' argument in that a claim
owner still has the ability to alter the flag value from the default where
an override is forced and cannot be changed.