Fix onBlockNotify not blocking updates to wilderness.
If a block in a claim notifies its surrounding blocks and one of those
blocks ends up in the wilderness, GP will now validate before allowing the
tracking to be updated. If the target position already contains tracking
for a user not trusted by the claim, the update will be blocked. This
prevents situations where a non-trusted player places a block at edge of a
claim that attempts to break a block inside a claim but cannot until a
trusted player accidentally causes an update to the block position which
allows it through.

Note: To see tracking in any block position, you need to have SpongeForge
installed on client and have necessary permissions to view tracking.
Simply press F3 and look at a block to check owner or notifier.

* Fix interact-entity flags ignoring players.
* Fix entity-damage flags not handling players properly.
* Add flag override support for interact-entity and entity-damage flags.