Add claim level feature.
Admins now have the ability to set options to restrict claim levels for
individual players or groups.

The new options are as follows :
  - griefprevention.min-claim-level
  - griefprevention.max-claim-level

If a player is in 2D mode, the claim will consume the entire space between
min and max level. Any claim found overlapping within the min and max levels while in
2D mode will prevent claim creation.
If a player is in 3D mode, the claim levels will be constraints. This
means a 3D claim can be created anywhere within the defined levels for the
player. Normal 3D overlap checks still apply.

* Added ClaimBlockSystem interface to API.
  - This new interface represents the claim block system used by GP.
  - The default is AREA which calculates each claim by x and z. Volume
    calculates x, y, and z.
* Add ChangeClaimEvent. This new event is fired for any changes to a
  claim. It currently supports Resize and Type.
* Moved ResizeClaimEvent to 'ChangeClaimEvent.Resize'.
* Renamed 'wilderness-cuboids' config option to 'claim-block-system' which
  now only accepts 'AREA' or 'VOLUME'.
* Removed all 3D restrictions in wilderness to support new level feature.
* Fixed /gpreload reloading PlayerData more than once.