Add blacklist feature. Closes #653
The blacklist feature gives admins the ability to control which item,
block, or entity should be ignored by GP's protection either globally or
on a per flag basis. To use, locate GP's global or a world config under
./config/griefprevention/worlds and edit the blacklist section to your
liking. The blacklist supports any valid item, block, or entity id as
well as wildcards '*' and '?'.