Add border-block-radius for extra protection around claims.
Many modded blocks, when placed next to claim border, can cause grief to
surrounding blocks. To solve this issue, set 'border-block-radius' to 1
or higher depending on how deep the block searches.
Note: The higher you set this setting, the more GP has to search for
claims. This can cause performance to degrade so use wisely.

Fixes #290

* Change config 'auto-claim-radius' to 'auto-chest-claim-block-radius'.
Note: If you had this disabled previously, make sure to disable the new
setting.
* Removed block place surrounding check.